Use read_path_frame() in read_ascii_data*()
Once the read/write functions used directly by the tracking code are reduced
to the bare essential lines required for handling the globals, I hope it'll be
easy to replace it all with the frame-buffer rewrite. Right now we have no
choice but to leave the code handling the globals in.