Should now be able to use setup.py on mac without hacks.
The problem was that gcc allows variables to be defined several times in
several separate files, and the linker merges them. This is called "common
mode". This is not allowed on mac, and ugly besides. So I added a compile
swithch '-fno-common', defined everything in globals.h extern and made sure
everything has one definition and one only.